cleanup invite & update logic
