--- web/src/components/forms/UserInvitationForm/graphql/inviteUser.graphql
@@ -1,5 +1,0 @@
-mutation InviteUser($input: InviteUserInput!) {
-    inviteUser(input: $input) {
-        id
-    }
-}
