+++ web/src/components/sidesheets/UserInvitationSidesheet/UserInvitationSidesheet.tsx
@@ -16,13 +16,31 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { Box, Heading, Text, SideSheet } from 'pouncejs';
-import InviteUserForm from 'Components/forms/UserInvitationForm';
+import { Box, Heading, Text, SideSheet, useSnackbar } from 'pouncejs';
 import React from 'react';
 import useSidesheet from 'Hooks/useSidesheet';
+import { extractErrorMessage } from 'Helpers/utils';
+import UserForm from 'Components/forms/UserForm';
+import { getOperationName } from '@apollo/client/utilities/graphql/getFromAST';
+import { ListUsersDocument } from 'Pages/Users';
+import { useInviteUser } from './graphql/inviteUser.generated';
+
+const initialValues = {
+  email: '',
+  familyName: '',
+  givenName: '',
+};
 
 const UserInvitationSidesheet: React.FC = () => {
   const { hideSidesheet } = useSidesheet();
+  const { pushSnackbar } = useSnackbar();
+  const [inviteUser, { error }] = useInviteUser({
+    onCompleted: data => {
+      hideSidesheet();
+      pushSnackbar({ variant: 'success', title: `Successfully invited ${data.inviteUser.email}` });
+    },
+    refetchQueries: [getOperationName(ListUsersDocument)],
+  });
 
   return (
     <SideSheet open onClose={hideSidesheet}>
@@ -34,13 +52,30 @@ const UserInvitationSidesheet: React.FC = () => {
           By inviting users to join your organization, they will receive an email with temporary
           credentials that they can use to sign in to the platform
         </Text>
-
-        <InviteUserForm onSuccess={hideSidesheet} />
+        <UserForm
+          initialValues={initialValues}
+          onSubmit={async values => {
+            await inviteUser({
+              variables: {
+                input: {
+                  email: values.email,
+                  familyName: values.familyName,
+                  givenName: values.givenName,
+                },
+              },
+            });
+          }}
+        />
         <Text size="small" color="grey300" textAlign="center" mt={6}>
           All users in the Open-Source version of Panther are admins in the system.
           <br />
           Role-based access is a feature available in the Enterprise version.
         </Text>
+        {error && (
+          <Text size="large" mt={6} color="red300">
+            {extractErrorMessage(error)}
+          </Text>
+        )}
       </Box>
     </SideSheet>
   );
