+++ web/src/components/sidesheets/UserInvitationSidesheet/graphql/inviteUser.generated.ts
@@ -28,12 +28,13 @@ export type InviteUserVariables = {
   input: Types.InviteUserInput;
 };
 
-export type InviteUser = { inviteUser?: Types.Maybe<Pick<Types.InviteUserResponse, 'id'>> };
+export type InviteUser = { inviteUser: Pick<Types.User, 'id' | 'email'> };
 
 export const InviteUserDocument = gql`
   mutation InviteUser($input: InviteUserInput!) {
     inviteUser(input: $input) {
       id
+      email
     }
   }
 `;
