+++ web/src/components/sidesheets/UserInvitationSidesheet/graphql/inviteUser.graphql
@@ -1,0 +1,6 @@
+mutation InviteUser($input: InviteUserInput!) {
+  inviteUser(input: $input) {
+    id
+    email
+  }
+}
